// src/app/enums/debuff.ts
export enum Debuff {
  BleedDamage = 'BleedDamage',
  PhysicalDamage = 'PhysicalDamage',
  GlobalCrit = 'GlobalCrit',
  SpellCrit = 'SpellCrit',
  SpellHit = 'SpellHit',
  SpellDamage = 'SpellDamage',
  Armor20 = 'Armor20',
  Armor5 = 'Armor5',
  AttackSpeedReduction = 'AttackSpeedReduction',
  AttackPowerReduction = 'AttackPowerReduction',
  HitReduction = 'HitReduction',
  PaladinJudgement = 'PaladinJudgement'
}