// src/app/enums/spec-name.ts
export enum SpecName {
  Balance = 'Balance',
  FeralDPS = 'FeralDPS',
  FeralTank = 'FeralTank',
  RestoDruid = 'ResoDruid',
  Arms = 'Arms',
  Fury = 'Fury',
  ProtWar = 'ProtWar',
  HolyPal = 'HolyPal',
  ProtPal = 'ProtPal',
  Retri = 'Retri',
  Assa = 'Assa',
  Combat = 'Combat',
  MM = 'MM',
  Surv = 'Surv',
  Disc = 'Disc',
  HolyPriest = 'HolyPriest',
  Shadow = 'Shadow',
  Affli = 'Affli',
  Demo = 'Demo',
  Elem = 'Elem',
  Ench = 'Ench',
  RestoShaman = 'RestoShaman',
  Blood = 'Blood',
  Frost = 'Frost',
  Unholy = 'Unholy',
  Fire = 'Fire',
  Arcane = 'Arcane',
}